+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## v1.0.1 - 2025-12-15
+- Start/Stop knoppen + aftelmechanisme (live countdown)
+- Systeemvak icoon (tray) met menu
+
 ## v1.0.0 - 2025-12-15
 - Eerste release: shutdown/restart plannen met kalender + seconden countdown
-